requirements.txt
python-dotenv==1.0.1
Faker==24.11.0
python-dateutil==2.9.0.post0
requests==2.31.0
tqdm==4.66.4
